Added support for pt-BR language
I'm uploading a Brazilian Portuguese version your lang file, because you've finally updated the Thaumic Wands mod on CurseForge.

Thank you for that and for saving me from having to recreate the entire mod just to add that memory leak fix.

And I'm also uploading a pt-BR lang file to Thaumic Wands as well. If you could take a look, would be appreciated.